% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benford-package.R
\docType{package}
\name{benford-package}
\alias{benford}
\alias{benford-package}
\title{benford: Implementação em GNU R da Lei de Benford ou Lei dos Números Anômalos}
\description{
Realiza alguns testes com a lei de Benford e gera gráficos.
}
\author{
\strong{Maintainer}: Marcelo Camacho \email{marcelocamacho.ufpa@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
