% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benford.R
\name{plot.benford}
\alias{plot.benford}
\title{Plota o gráfico para a Lei de Benford}
\usage{
\method{plot}{benford}(x)
}
\arguments{
\item{x}{resultado da função \code{\link{benford}}}
}
\description{
Gráfico de barras
}
\details{
Utiliza o pacote ggplot2
}
