% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benford.R
\name{benford}
\alias{benford}
\title{Lei de benford}
\usage{
benford(x)
}
\arguments{
\item{x}{Vetor de inteiros}
}
\value{
uma lista com estatística, valor-p,
frequências observadas e esperadas.
}
\description{
Esta função lê um vetor de números inteiros e testa se
vale a lei de benford para esses números.
}
\details{
Histórico sobre a Lei de Benford.
}
